LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/auto-005-expected.txt

1

FAIL contain-intrinsic-size: auto assert_equals: expected 50 but got 1
